-Wsign-compare

+++ cpu_compute0.cpp
@@ -552,11 +552,11 @@ l_end_pixel: ;
    if(filter == EVFILT_SIGNAL) { /* signal */
     fprintf(stderr, " ! got %s\n", strsignal(id));
     f_exit_sig = true;
-   }else if(id == x11_fd) { /* x11 event */
+   }else if((int)id == x11_fd) { /* x11 event */
     aux_xcb_ev_func(&xcb_ctx);
     f_win_close  = xcb_ctx.f_window_should_close;
     f_win_expose = xcb_ctx.f_window_expose;
-   }else if(id == dri_fd) { /* DRM event */
+   }else if((int)id == dri_fd) { /* DRM event */
 	struct  aux_drm_event_crtc_sq vev[4];
 
     /* drm events are written in full */
